+++ Functions/Describe.ps1
@@ -33,22 +33,22 @@ function Add-Numbers($a, $b) {
 Describe "Add-Numbers" {
     It "adds positive numbers" {
         $sum = Add-Numbers 2 3
-        $sum | Should Be 5
+        $sum | Should -Be 5
     }
 
     It "adds negative numbers" {
         $sum = Add-Numbers (-2) (-2)
-        $sum | Should Be (-4)
+        $sum | Should -Be (-4)
     }
 
     It "adds one negative number to positive number" {
         $sum = Add-Numbers (-2) 2
-        $sum | Should Be 0
+        $sum | Should -Be 0
     }
 
     It "concatenates strings if given strings" {
         $sum = Add-Numbers two three
-        $sum | Should Be "twothree"
+        $sum | Should -Be "twothree"
     }
 }
 
@@ -179,7 +179,7 @@ function DescribeImpl {
     }
     catch
     {
-        $firstStackTraceLine = $_.InvocationInfo.PositionMessage.Trim() -split '\r?\n' | & $SafeCommands['Select-Object'] -First 1
+        $firstStackTraceLine = $_.InvocationInfo.PositionMessage.Trim() -split "$([System.Environment]::NewLine)" | & $SafeCommands['Select-Object'] -First 1
         $Pester.AddTestResult("Error occurred in $CommandUsed block", "Failed", $null, $_.Exception.Message, $firstStackTraceLine, $null, $null, $_)
         if ($null -ne $TestOutputBlock)
         {