+++ Functions/It.Tests.ps1
@@ -7,30 +7,30 @@ InModuleScope Pester {
 
         It 'Throws an error if you fail to pass in a test block' {
             $scriptBlock = { ItImpl -Pester $testState 'Some Name' }
-            $scriptBlock | Should Throw 'No test script block is provided. (Have you put the open curly brace on the next line?)'
+            $scriptBlock | Should -Throw 'No test script block is provided. (Have you put the open curly brace on the next line?)'
         }
 
         It 'Does not throw an error if It is passed a script block, and adds a successful test result.' {
             $scriptBlock = { ItImpl -Pester $testState 'Enters an It block inside a Describe' { } }
-            $scriptBlock | Should Not Throw
+            $scriptBlock | Should -Not -Throw
 
-            $testState.TestResult[-1].Passed | Should Be $true
-            $testState.TestResult[-1].ParameterizedSuiteName | Should BeNullOrEmpty
+            $testState.TestResult[-1].Passed | Should -Be $true
+            $testState.TestResult[-1].ParameterizedSuiteName | Should -BeNullOrEmpty
         }
 
         It 'Does not throw an error if the -Pending switch is used, and no script block is passed' {
             $scriptBlock = { ItImpl -Pester $testState 'Some Name' -Pending }
-            $scriptBlock | Should Not Throw
+            $scriptBlock | Should -Not -Throw
         }
 
         It 'Does not throw an error if the -Skip switch is used, and no script block is passed' {
             $scriptBlock = { ItImpl -Pester $testState 'Some Name' -Skip }
-            $scriptBlock | Should Not Throw
+            $scriptBlock | Should -Not -Throw
         }
 
         It 'Does not throw an error if the -Ignore switch is used, and no script block is passed' {
             $scriptBlock = { ItImpl -Pester $testState 'Some Name' -Ignore }
-            $scriptBlock | Should Not Throw
+            $scriptBlock | Should -Not -Throw
         }
 
         It 'Creates a pending test for an empty (whitespace and comments only) script block' {
@@ -43,8 +43,8 @@ InModuleScope Pester {
                 #>
             }
 
-            { ItImpl -Pester $testState 'Some Name' $scriptBlock } | Should Not Throw
-            $testState.TestResult[-1].Result | Should Be 'Pending'
+            { ItImpl -Pester $testState 'Some Name' $scriptBlock } | Should -Not -Throw
+            $testState.TestResult[-1].Result | Should -Be 'Pending'
         }
 
         It 'Adds a failed test if the script block throws an exception' {
@@ -54,10 +54,10 @@ InModuleScope Pester {
                 }
             }
 
-            $scriptBlock | Should Not Throw
-            $testState.TestResult[-1].Passed | Should Be $false
-            $testState.TestResult[-1].ParameterizedSuiteName | Should BeNullOrEmpty
-            $testState.TestResult[-1].FailureMessage | Should Be 'I am a failed test'
+            $scriptBlock | Should -Not -Throw
+            $testState.TestResult[-1].Passed | Should -Be $false
+            $testState.TestResult[-1].ParameterizedSuiteName | Should -BeNullOrEmpty
+            $testState.TestResult[-1].FailureMessage | Should -Be 'I am a failed test'
         }
 
         $script:counterNameThatIsReallyUnlikelyToConflictWithAnything = 0
@@ -69,7 +69,7 @@ InModuleScope Pester {
             ItImpl -Pester $testState 'Does something' -OutputScriptBlock $outputBlock { }
             ItImpl -Pester $testState 'Does something' -OutputScriptBlock $outputBlock { }
 
-            $script:counterNameThatIsReallyUnlikelyToConflictWithAnything | Should Be 3
+            $script:counterNameThatIsReallyUnlikelyToConflictWithAnything | Should -Be 3
         }
 
         Remove-Variable -Scope Script -Name counterNameThatIsReallyUnlikelyToConflictWithAnything
@@ -91,13 +91,13 @@ InModuleScope Pester {
             ItImpl -Pester $testState -Name $suiteName -TestCases $cases {
                 param ($a, $b, $expectedResult)
 
-                ($a + $b) | Should Be $expectedResult
+                ($a + $b) | Should -Be $expectedResult
             }
 
             It 'Creates test result records with the ParameterizedSuiteName property set' {
                 for ($i = -1; $i -ge -4; $i--)
                 {
-                    $testState.TestResult[$i].ParameterizedSuiteName | Should Be $suiteName
+                    $testState.TestResult[$i].ParameterizedSuiteName | Should -Be $suiteName
                 }
             }
 
@@ -105,15 +105,15 @@ InModuleScope Pester {
                 for ($i = -1; $i -ge -4; $i--)
                 {
                     $expectedName = "Adds $($cases[$i]['a']) and $($cases[$i]['b']) to get $($cases[$i]['expectedResult']).  <Bogus> is not a parameter."
-                    $testState.TestResult[$i].Name | Should Be $expectedName
+                    $testState.TestResult[$i].Name | Should -Be $expectedName
                 }
             }
 
             It 'Logs the proper successes and failures' {
-                $testState.TestResult[-1].Passed | Should Be $false
+                $testState.TestResult[-1].Passed | Should -Be $false
                 for ($i = -2; $i -ge -4; $i--)
                 {
-                    $testState.TestResult[$i].Passed | Should Be $true
+                    $testState.TestResult[$i].Passed | Should -Be $true
                 }
             }
         }
@@ -139,20 +139,20 @@ InModuleScope Pester {
 
         It 'Reports keys and values in the same order as the param block' {
             ($dictionary.Keys -join ',') |
-            Should Be '1,c,0,z,a,Something.Really/Weird '
+            Should -Be '1,c,0,z,a,Something.Really/Weird '
 
             ($dictionary.Values -join ',') |
-            Should Be 'One,C,Zero,Z,A,Weird'
+            Should -Be 'One,C,Zero,Z,A,Weird'
         }
     }
 
     Describe 'Remove-Comments' {
         It 'Removes single line comments' {
-            Remove-Comments -Text 'code #comment' | Should Be 'code '
+            Remove-Comments -Text 'code #comment' | Should -Be 'code '
         }
         It 'Removes multi line comments' {
             Remove-Comments -Text 'code <#comment
-            comment#> code' | Should Be 'code  code'
+            comment#> code' | Should -Be 'code  code'
         }
     }
 }