+++ Functions/It.ps1
@@ -54,22 +54,22 @@ function Add-Numbers($a, $b) {
 Describe "Add-Numbers" {
     It "adds positive numbers" {
         $sum = Add-Numbers 2 3
-        $sum | Should Be 5
+        $sum | Should -Be 5
     }
 
     It "adds negative numbers" {
         $sum = Add-Numbers (-2) (-2)
-        $sum | Should Be (-4)
+        $sum | Should -Be (-4)
     }
 
     It "adds one negative number to positive number" {
         $sum = Add-Numbers (-2) 2
-        $sum | Should Be 0
+        $sum | Should -Be 0
     }
 
     It "concatenates strings if given strings" {
         $sum = Add-Numbers two three
-        $sum | Should Be "twothree"
+        $sum | Should -Be "twothree"
     }
 }
 
@@ -90,7 +90,7 @@ Describe "Add-Numbers" {
         param ($a, $b, $expectedResult)
 
         $sum = Add-Numbers $a $b
-        $sum | Should Be $expectedResult
+        $sum | Should -Be $expectedResult
     }
 }
 
@@ -280,7 +280,7 @@ function Invoke-Test
                 $pester.LeaveTest()
             }
 
-            $result = ConvertTo-PesterResult -ErrorRecord $errorRecord
+            $result = ConvertTo-PesterResult -Name $Name -ErrorRecord $errorRecord
             $orderedParameters = Get-OrderedParameterDictionary -ScriptBlock $ScriptBlock -Dictionary $Parameters
             $Pester.AddTestResult( $result.name, $result.Result, $null, $result.FailureMessage, $result.StackTrace, $ParameterizedSuiteName, $orderedParameters, $result.ErrorRecord )
             & $SafeCommands['Write-Progress'] -Activity "Running test '$Name'" -Completed -Status Processing